collector/requirements.txt
feedparser==6.0.11
chromadb==0.5.11
httpx==0.27.2
python-dotenv==1.0.1
